Rename yaml_eigen_extension to yaml_extension
